💄 Style: auth layout ui 수정, ♻️ Refactor: layout 리팩토링
